Projectile Collision
-Added Projectile Collision (Every tile has a solid boolean for the
player, and a projectileSolid boolean for the projectile. Example: Water
is solid for the play but not the projectile
-Optimizations! No longer clearing the bufferedimage to be displayed
which increased my FPS by about 15%
-Added a particle spawner that an object can use to spawn lots of
particles

Warning: Do not shoot off the screen right now with a projectile or it
will crash the game

FYI: This will probably  be my last commit before vacation! But we will
plan the game and when I get back we will start development again :D