My weekly update!
Have been holding off on committing this for awhile, wanted to iron out
some bugs, fix some things. The minimap hurts performance by about 15%,
will be looking into ways in the future to make it better, but for now
it will stay as is :) Found one but that is pretty game breaking, and
that is running the game on fullscreen when the resolution does not fit
your monitor. Your mouse precision will be way off!:P Settings should be
touched up soon to save your previous setting, and there are a few minor
bugs with it, one is posted on the Trello. Other then that I will be
continuing work, I will probably do the projectile class clean up soon
:) Anyways on to what I added:

-Added right click interaction (A state that extends LevelState can set
its own right click interactions)
-Added cursor changing options on the fly (Used in conjunction with
hovering your mouse cursor over an object you can interact with)
-Cleaned up the Tile class (Combined many tiles into one big mother
class ie InterchangeableFloorTile)
-Added a bunch of new Sprites and HUD stuff
-Fixed a bug where dead slime corpses would absorb projectiles
-Added a new mouse cursor image! (Exciting stuff)
-Added a HUD with a working Health bar which will scale when you take
damage or become healed. Also added a minimap, might need some
performance tweaking and updated textures eventually, but it is not
important right now
-Small clean up in other classes and slight render loop adjustment to
render a couple less tiles at a time (to be precise 40 in 1920x1080 :+1:
)

Probably won't be pushing an update for another week or so. Got some
important stuff coming up in school :)