Polish for the Public!
-Added a game icon
-Added a game project name

I am letting people on my Skype test the game soon, and I will be giving
them a survey. It will cover their hardware and their FPS and UPS. So I
can tell if I need to make optimizations or if the game is in a decent
state to move onto bigger gameplay things like naval combat!