Combat Imrpoved #2! (And other stuff)
-Added melee and range form swapping, this will allow the player to
melee and use ranged attacks. However your gun will not reload when in
melee form. (The mousewheel changes forms, this led to me adding a
MouseWheelListener to the MoustMaster class)

-Added a reload bar when in ranged form. When you shoot a bar will pop
up and will get smaller and when it is gone, that means the player is
done reloading. Keeping in mind you must be in ranged form for the
cooldown to go down.

---------------Other Changes--------------------------

-Added a main menu tilemap, so the main menu is more fancy and looks
good :) Will end up making the other menu's look better in the future,
but not for now, because I am going to go work on mechanics and mobs and
such :)

-Added the changing form controls to the control menu

-Took the gameplay states that the player will be playing on. And put
them in a seperate package called gameplaystates so they are no longer
in the gamestatemanager package, now just main menu and settings and
inventory state and such will be put it in the gamestatemanager class,
no longer gameplaystates :)
(Keeps it looking neat)