Huge gameplay update
Going to start working on this project more, did these updates in the
past three days :)

-Fixed Projectiles deconstructing
-Made a chat bar texture
-Added TextBoxMaster class (Only works when the player is in a
level/gameplay state)
-Made an enemy package (moved slime to it)
-Made an NPC package (moved king and wiseman into it)
-Made an NPCAnimate class that all of the NPC's can use
-Added a new tilemap function, can replace tiles on the fly (will be
used to lock the player in by replacing the floor with a solid wall)
- Added a visual flicker when invincible (when the player gets hit he
will flicker to showcase he got hurt and how long he is invincible for)
-Added a string type to tiles to interchangeable tiles can be different
(used for slimes as now they only spawn on grass tiles)
-moved spawn chance to the levelstate instead of SlimeSpawner
-Made SlimeSpawner work properly, the player will no longer see slimes
spawn and slimes only spawn on grass (not in solids anymore :D) I will
update this in the future so only so many slimes will spawn and some
will despawn after awhile and such
-Added a visual for doing a melee attack (might replace texture)
-Balanced slimes by making higher level slimes move faster
-Made the slime spawner, spawn level 1, 2, and 3 slimes
-Renamed quit to Save and Quit (on the pause menu)
-Wrote an algorithm that executes when the player hits 'f' to talk to
someone. It will determine the closest NPC and if the NPC is close
enough it will open an TextBox with the NPC's dialog in it. Entities now
have a description which is the dialog they speak
-Fixed the HUD class requiring two updates to properly display
information
-Added a new NPC, King and Wiseman (wiseman is in the house)
-Fixed a visual bug involving pauses and moving (not perfect yet but
better)
-Added working xp bar with level cap functionality
-Added persistent character level and xp
-Added a player method that should be called when changing an instance,
it saves the player before the player gets reinitialized in the new
instance

Will be working on the leveling, adding more enemies, NPC's, bug fixing,
and working on a new system for loading Sprites instead of loading them
all at once.