FINAL UPDATE - MOVING ON
Added bad shit movement with no animations. I decided to not update this
game anymore due to all of the spaghetti code, that I wrote over a
weekend due to this game being made for a game jam. This engine is not
very good as I made it when I was new to programming, it handles
resources inefficiently, is not very fast at all (it is Java2D graphics
too). If I decide to make an actual polished game in the future I will
write my own engine in C++ using DirectX or something, or I will just
use something like Unity 3D.

I should have probably stopped working on this project after the game
jam, because it was low quality and made in a weekend.

There is many bugs with the ship movement that I added, but I thought
since it is a pirate game, I needed to add it!

Thanks to anyone who played the game, or continued to try out my new
updates!