Crash Fixes and performance increase
- Stopped the tilemap from updating all of the tiles on the map, so that
means that the algorithm to update the map which was barely used for
anything, is know no longer in the code. If n = # of tiles. Then the
update tilemap algorithm had a complexity of O(n). Every game update, so
60 times a second. Considering the base map had over 65000 tiles, that
saves a lot of random iterating. This was in the game, because I made
this game engine in grade 12, so I was still a very inexperienced
programmer. I plan to rewrite the game eventually in C++ with OpenGL or
DirectX.

- Fixed the game from crashing due to slimes being spawned on the edge
of the map or projectiles reaching the end of the map.